refactor: Implement Domain-Driven Design Modularization
- Decoupled `ICATEngine`, `icat_schema.py`, and `snc_utils.py` out of `core/engine/` to the new deterministic `adam_finance/` context.
- Extracted `calculate_var` and `calculate_cvar` functions from `QuantitativeRiskAgent` into the pure, deterministic `adam_finance/math.py` module.
- Introduced Interface Segregation by defining `FinancialModelProtocol` inside the new `adam_interfaces/interfaces.py` layer, and made the `QuantitativeRiskAgent` implement it.
- Executed prompt eviction by migrating the system prompt out of `core/agents/specialized/credit_risk_controller_agent.py` to `assets/credit_risk_controller_prompt.md`.
- Wrote expansive architectural plans in `docs/ARCHITECTURE_PLAN.md` and synthesized `README.md` files for the extracted bounded contexts to support 'Search Before Summarize' LLM context preservation.
- Updated all import pathways accordingly, confirmed tests pass locally.
